Reference

Reference

Benchmarks

Newtman.Easom — Type.

Easom

An unconstrained implementation of the 2-dimensional Easom function defined as:

\[f(\mathbf{x}) = -\cos{(x_1)} \cos{(x_2)} \exp{[-(x_1 - \pi)^2 - (x_2 - \pi)^2]}\]

where $x_1$ and $x_2$ refer to the first and second element of the input vector $\mathbf{x}$.

source

Newtman.Sphere — Type.

Sphere

An unconstrained implementation of the Sphere function defined as:

\[f(\mathbf{x}) = \sum_{i=1}^{d} x_i^2\]

where $d$ is the dimension of the input vector $\mathbf{x}$.

source

Newtman.TestFunctions — Type.

TestFunctions

Abstract supertype for all benchmark functions.

source

Newtman.Unconstrained — Type.

Unconstrained

Abstract supertype for all unconstrained benchmark functions.

source

Algorithms

Newtman.PSO — Type.

PSO

PSO is the type associated with the implementation for the Particle Swarm Optimization with momentum. See Algorithms for more information.

source

Newtman.PSO — Method.

PSO(f::Function, population::AbstractArray, k_max::Int;
    w=0.9, c1=2.0, c2=2.0) -> OptimizationResults

Method that implements PSO for a function f of type Function, or of type TestFunctions. Returns an OptimizationResults type with information relevant to the run executed, see OptimizationResults.

Arguments

population: can be any AbstractArray that contains Particle

instances, but it is expected to be generated by Population.

k_max: number of maximum iterations until "convergence" of the algorithm.

Keyword arguments

It is recommended to use the default values provided.

w: value that controls how much of the initial velocity is retained, i.e.

an inertia term. This values decays linearly over each iteration until it reaches the default miminum value of 0.4.

c1: balance between the influence of the individual's knowledge, i.e. the

best inidividual solution so far.

c2: balance between the influence of the population's knowledge, i.e. the

best global solution so far.

Examples

For Functions

using Newtman

# Define the Sphere function
f_sphere(x) = sum(x .^ 2)

# Implement PSO for a 3-dimensional Sphere function, with
# 10000 iterations and 30 particles in the population.
val = PSO(f_sphere, Population(30, 3, -15.0, 15.0), 10000)

For TestFunctions

using Newtman

# Implement PSO for a 3-dimensional Sphere function, with
# 10000 iterations and 25 particles in the population.
val = PSO(Sphere(), Population(25, 3, -15.0, 15.0), 10000)

source

Newtman._clip_positions_velocities! — Method.

Apply boundary conditions to both position and velocity for every Particle type object P.

source

Newtman._create_rng — Method.

Returns num_rngs number of statistically independent pseudo-random number generators by creating a master RNG and drawing numbers from it, these numbers will serve as seeds for the number of RNG's returned in the form of an AbstractArray.

source

Newtman._mean_std_results — Method.

Receives an array of values and a num_max number, and returns the mean and standard deviation of them.

source

Newtman._weight_decay — Method.

Compute the corresponding weight decay depending the maximum number of iterations and the initial value for it.

source

Newtman.Metaheuristic — Type.

Metaheuristic

Abstract type for metaheuristic algorithms, this makes a clear distinction between different classifications of metaheuristic algorithms.

source

Newtman.OptimizationResults — Type.

OptimizationResults{T, U, V, W}

Type that formats the output of Solver to get better information from it.

Fields

x::T: Stores the solution array from the solver, i.e. the solution that minimizes the cost function.

min::U: Stores the value obtained from evaluating the cost function with x, i.e. the minima found.

impl::AbstractString: Stores the name of the Solver used, i.e. the name or identifier of the optimization algorithm.

iterations::Integer: Stores the number of maximum iterations that the solver was run.

source

Newtman.OptimizationResultsParallel — Type.

OptimizationResultsParallel{T, U}

Type that formats the output of an embarrassingly parallel implementation of Solver to get better information from it.

Fields

x::T: Stores the solution array from the solver, i.e. the solution that minimizes the cost function. This is the mean value obtained from all the independent runs that were carried out.

err::U: Stores the standard deviation from the x array.

min::U: Stores the value obtained from evaluating the cost function with x, i.e. the minima found.

impl::AbstractString: Stores the name of the Solver used, i.e. the name or identifier of the optimization algorithm.

iterations::Integer: Stores the number of maximum iterations that the solver was run.

runs::Integer: Stores the number of independent runs that were executed.

source

Newtman.PopulationBase — Type.

PopulationBase

Type for population-based algorithms that employ Population, i.e. subroutines that mutate an array of possible candidates in-place. An example of this type is PSO.

source

Newtman.Solver — Type.

Solver

Abstract super-type for every algorithm implementation.

source

Population

Newtman.Particle — Method.

Particle(x::T, v::T, x_best::T, a::V, b::V)
    where {T<:AbstractArray, V<:AbstractFloat}

A type that can hold information about current position, current velocity, the best candidate to a solution, as well as defining the bounds. The dimensions of the Particle are inferred from the length of the arrays.

Arguments

x: Array that holds the positions of possible solutions.

v: Array that holds velocities related to x.

x_best: An element of x that determines the best position for the particle.

a: lower bound for x

b: upper bound for v

Example

p = Particle(zeros(3), rand(3), zeros(3), -1.0, 1.0)

source

Newtman.Particle — Method.

Particle(a::T, b::T, n::V)
    where {T<:AbstractFloat, V<:Int}

Particle that can be created randomly using the bounds and the dimension needed.

Arguments

a: lower bound for x

b: upper bound for v

n: dimension for x, v, and x_best.

Example

p = Particle(-1.0, 1.0, 3)

source

Newtman.Population — Method.

Population(num_particles::T, dim::T, a::V, b::V)
    where {T<:Int, V<:AbstractFloat} -> Vector{Particle}(undef, num_particles)

An array of Particles where each of them are bounded and are given a dimension. This is essentially a multi-dimensional array. It makes handling Particles much easier.

Arguments

num_particles: Number of particles in the Population.

dim: Dimension for every Particle.

a: Lower bound for every Particle, this is shared across every instance.

b: Upper bound for every Particle, this is shared across every instance.

Example

pop = Population(35, 4, -1.0, 1.0)

source

Newtman.Population — Method.

Population(dim::T, a::V, b::V)
    where {T<:Int, V<:AbstractFloat} -> Vector{Particle}(undef, num_particles)

If num_particles is not provided, it defaults to 5 Particles in the Population.

Arguments

dim: Dimension for every Particle.

a: Lower bound for every Particle, this is shared across every instance.

b: Upper bound for every Particle, this is shared across every instance.

Example

pop = Population(4, -1.0, 1.0) # The same as Population(5, 4, -1.0, 1.0)

source

Newtman.Individual — Type.

Individual

Abstract super-type for types that contain their own information.

source